third_party/blink/web_tests/platform/win11-arm64/http/tests/security/contentSecurityPolicy/sandbox-invalid-header-expected.txt

CONSOLE MESSAGE: Script executed in iframe.
